k8s_mcp/requirements.txt
flask==2.3.3
requests==2.31.0
python-dotenv==1.0.0
markdown==3.4.4
beautifulsoup4==4.12.2
gunicorn==21.2.0
pyyaml==6.0.1
kubernetes==28.1.0
flask-cors==4.0.0 